Add Lightning Bolt.
Remove cast time from SpellCastingTime constructor and let inherited
classes set the variable.

Update crit chance for MagicAttackTable to use unsigned instead of
double.

+++ CombatRoll/AttackTables/MagicAttackTable.cpp
@@ -17,7 +17,7 @@ MagicAttackTable::MagicAttackTable(Mechanics* mechanics, Random* random, const i
     update_target_resistance(target_res);
 }
 
-int MagicAttackTable::get_hit_outcome(const unsigned roll, const double crit_chance) const
+int MagicAttackTable::get_hit_outcome(const unsigned roll, const unsigned crit_chance) const
 {
     if (roll < miss_range)
         return MagicAttackResult::MISS;
