+++ CombatRoll/AttackTables/MagicAttackTable.h
@@ -14,7 +14,7 @@ public:
     MagicAttackTable(Mechanics* mechanics, Random* random, const int clvl, const unsigned spell_hit, const int target_res);
     ~MagicAttackTable() = default;
 
-    int get_hit_outcome(const unsigned roll, const double crit_chance) const;
+    int get_hit_outcome(const unsigned roll, const unsigned crit_chance) const;
     int get_resist_outcome(const unsigned roll) const;
     void update_miss_chance(const int clvl, const unsigned spell_hit);
     void update_target_resistance(const int target_resistance);
